+++ CHANGELOG.md
@@ -115,3 +115,4 @@
   - Minor bug fixes #143
 - v0.5.2
   - Add save options in Dymola #98
+  - Add feature to postprocess mat results within the simulate function to avoid memory errors in large studies